+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Chore & Maintenance
 
-- `[expect]` Expose default matchers ([#11932](https://github.com/facebook/jest/pull/11932))
+- `[expect]` Export default matchers ([#11932](https://github.com/facebook/jest/pull/11932))
 
 ### Performance